% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_FleschKincaidGradeLevel.R
\name{compute_FleschKincaidGradeLevel}
\alias{compute_FleschKincaidGradeLevel}
\title{Computes Flesch-Kincaid Grade Level on Czech text.}
\usage{
compute_FleschKincaidGradeLevel(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a numeric value
}
\description{
Takes a conll-u data frame of a Czech text and computes the Flesch-Kincaid Grade Level.
}
\seealso{
Bendová, K., & Cinková, S. (2021). Adaptation of Classic Readability Metrics to Czech. In K. Ekštein, F. Pártl, & M. Konopík (Eds.), Lecture Notes in Artificial Intelligence, 24th international Conference on Text, Speech and Dialogue (Vol. 12848, pp. 159–171). Springer.

Kincaid, J. P., Fishburne, R. P., Rogers, R. L., Chissom, B. S., & BRANCH, N. T. T. C. M. T. R. (1975). Derivation of New Readability Formulas (Automated Readability Index, Fog Count and Flesch Reading Ease Formula) for Navy Enlisted Personnel. Defense Technical Information Center. https://books.google.cz/books?id=7Z7ENwAACAAJ

\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{function udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}

create_syllablecount_per_token_df

count_words_in_conllu

filter_out_multiwords
}
