% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_average_sentence_length.R
\name{count_average_sentence_length}
\alias{count_average_sentence_length}
\title{Computes mean number of tokens per sentence}
\usage{
count_average_sentence_length(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a numeric value
}
\description{
Takes a conll-u data frame, removes punctuation rows with reconstructed full forms to actually spotted contracted forms
}
\seealso{
\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{function udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}

count_sentences_in_conllu

count_words_in_conllu

filter_out_multiwords
}
