% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_average_token_length_in_syllables.R
\name{count_average_token_length_in_syllables}
\alias{count_average_token_length_in_syllables}
\title{Computes mean number of syllables per token}
\usage{
count_average_token_length_in_syllables(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a numeric value
}
\description{
Takes a conll-u data frame, removes punctuation rows with reconstructed full forms to actually spotted contracted forms, counts the remaining tokens and syllables in each token, and divides the total of syllables by the total of tokens.
}
\examples{
\dontrun{
The example document contains 5 non-punctuation tokens: 

Abys (2 syllables), to (1 syllable), etc., 10 syllables in total.
 
The mean syllable count per token is 10/5 = 2.  

create_syllablecount_per_token_df(two_multitokens)

count_average_token_length_in_syllables(two_multitokens)
}
}
\seealso{
\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{function udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}

create_syllablecount_per_token_df

count_words_in_conllu

filter_out_multiwords
}
