% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_czech_syllables.R
\name{count_czech_syllables}
\alias{count_czech_syllables}
\title{Count Czech Syllables}
\usage{
count_czech_syllables(token)
}
\arguments{
\item{token}{character string, character vector. Must be encoded in UTF-8.}
}
\value{
numeric value
}
\description{
Takes a Czech word and counts syllables.
}
\details{
This function counts syllables in a Czech word. The text must be encoded in UTF-8. The syllable is mostly defined either as a single vowel or the diphthongs "au" and "ou". Since not all sequences of "au" and "ou" are diphthongs, the function uses a built-in lexicon of stemmed words starting with "u". Besides, the function checks for frequent compounds ending with "o", such as "samo" (e.g. "samouk"), "celo", "velko", "rychlo",  and a few others. Other cases considered are the Czech syllabic "l" or "r"; e.g. in "mrkl" (he blinked), and the German diphthong "ei". On the other hand, the German "ie" remains two syllables, as it is  pronounced in Czech (loan) words (e.g. "Marie", "historie"). Non-syllabic tokens except punctuation, such as digits or symbols, count as monosyllabic.

This lexicon was extracted from the Czech National Corpus (Araneum Bohemicum Maximum, Capek complete, and SYN Version 11), stemmed, and overtly manually cleaned. The searches targeted lemmas starting with "u" and were complemented by searches for lemmas where "u" was preceded by potential prefixes e.g.("na, za, pro, po, do"). If your Czech text has been stripped of diacritics, the lexicon-based de-diphtongization will be less sensitive. Not completely eliminated though, since the corpus returned a non-trivial amount of diacritics-less matches, and these were not deliberately removed.
}
\examples{
\dontrun{
count_czech_syllables("nauka") ## should have 3 syllables: found in lexicon

count_czech_syllables("pavouk") ## should have 2 syllables: ou is a diphthong

count_czech_syllables("polousazeni") ## should have 6 syllables: ou is not 

a diphthong, since the word is a compound with the first member listed
}

}
\author{
Silvie Cinkova
}
