% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sentences_in_conllu.R
\name{count_sentences_in_conllu}
\alias{count_sentences_in_conllu}
\title{Counts sentences in a conll-u file converted to data frame.}
\usage{
count_sentences_in_conllu(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a numeric value
}
\description{
Takes all unique combinations of document ID, paragraph ID, and sentence ID and counts them. It does not rely on the sentence IDs incrementing continuously. Hence no need to re-index the sentence ids when you remove some sentences before counting. It considers one data frame one document and gives no warning when the inputted data frame contains more than one document ID.
}
\examples{
\dontrun{
count_sentences_in_conllu(text_Vese)

Counts sentences in a document with 19 sentences in two paragraphs. 
}

}
\seealso{
\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}
}
