% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words_in_conllu.R
\name{count_words_in_conllu}
\alias{count_words_in_conllu}
\title{Counts tokens in a conll-u file converted to data frame.}
\usage{
count_words_in_conllu(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a numeric value
}
\description{
Takes all rows that correspond to actually spotted tokens (i.e., it ignores reconstructed full form components of contracted forms)
}
\examples{
count_words_in_conllu(itsgood) ## Counts tokens in a document with 3 tokens. 

}
\seealso{
filter_out_multiwords

\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}
}
