% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_multiwords.R
\name{filter_out_multiwords}
\alias{filter_out_multiwords}
\title{Removes reconstructed forms from multiword tokens in conll-u.}
\usage{
filter_out_multiwords(conlludf)
}
\arguments{
\item{conlludf}{a conll-u object converted to data frame with udpipe::udpipe_read_conllu()}
}
\value{
a data frame (conll-u data frame)
}
\description{
The conll-u data frame contains represents each spotted token with one row. Token ID of contracted forms (e.g. "it's") contains "-" separating the id of its first and last component (e.g. 7-8). These components (here "it" and "is") occur afterwards with regular token IDs (here 7 and 8). When counting tokens, you usually do not want these components but only the actually spotted contraction.
}
\examples{
\dontrun{
#' print(itsgood)
filter_out_multiwords(itsgood)
} 

}
\seealso{
\href{https://rdocumentation.org/packages/udpipe/versions/0.8.11/topics/udpipe_read_conllu}{function udpipe_read_conllu in Wijffels J (2023). \emph{udpipe: Tokenization, Parts of Speech Tagging, Lemmatization and Dependency Parsing with the 'UDPipe' 'NLP' Toolkit}. R package version 0.8.11,}
}
