% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadabilityCzech-package.R
\docType{package}
\name{ReadabilityCzech-package}
\alias{ReadabilityCzech}
\alias{ReadabilityCzech-package}
\title{ReadabilityCzech: What the Package Does (One Line, Title Case)}
\description{
Implementation of two classic readability formulas (Flesch Reading Ease, Flesch-Kincaid Grade Level) to Czech.
}
\author{
\strong{Maintainer}: Silvie Cinkova \email{cinkova@ufal.mff.cuni.cz} (\href{https://orcid.org/0000-0003-4526-3915}{ORCID})

}
\keyword{internal}
